pyproject.toml
[tool.poetry]
name = "rag-pipeline"
version = "0.1.0"
description = "RAG Pipeline with LangChain and Streamlit"
authors = ["dusvyat"]

[tool.poetry.dependencies]
python = ">=3.9.0,!=3.9.7,<3.13"
langchain = "^0.2.15"
streamlit = "^1.38.0"
unstructured = "^0.15.9"
python-docx = "^1.1.2"
pypdf = "^4.3.1"
chromadb = "^0.5.5"
openai = "^1.43.0"
langchain-community = "^0.2.15"
python-dotenv = "^1.0.1"
tiktoken = "^0.7.0"
langchain-aws = "^0.1.17"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
sqlalchemy = "^2.0.32"
passlib = "^1.7.4"
python-jose = "^3.3.0"
pydantic = "^2.8.2"
python-multipart = "^0.0.9"
requests = "^2.32.3"
pyjwt = "^2.9.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"